Organize Fission CLI release modules
